Implement ViewModel for ProductsPage.
The powerful concept of ViewModels comes from MVVM architectural design pattern. The change introduced via this commit shows the true strength of using such approach. ViewModel now has the responsibility to fetch the data, and the UI reacts on ViewModel's property change. This way, the same ProductsViewModel can be used for both ProductsPage and any other part of the application which displays user's products.